requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
textblob>=0.17.1
nltk>=3.8.1
spacy>=3.7.2